+++ apps/backend/src/app/api/latest/auth/oauth/callback/[provider_id]/route.tsx
@@ -165,11 +165,6 @@ const handler = createSmartRouteHandler({
         if (!user) {
           throw new StackAssertionError("User not found");
         }
-
-        const account = user.projectUserOAuthAccounts.find((a) => a.configOAuthProviderId === provider.id);
-        if (account && account.providerAccountId !== userInfo.accountId) {
-          throw new KnownErrors.UserAlreadyConnectedToAnotherOAuthConnection();
-        }
       }
 
       const oauthRequest = new OAuthRequest({
@@ -403,6 +398,7 @@ const handler = createSmartRouteHandler({
       } catch (error) {
         if (error instanceof InvalidClientError) {
           if (error.message.includes("redirect_uri") || error.message.includes("redirectUri")) {
+            console.log("User is trying to authorize OAuth with an invalid redirect URI", error, { redirectUri: oauthRequest.query?.redirect_uri, clientId: oauthRequest.query?.client_id });
             throw new KnownErrors.RedirectUrlNotWhitelisted();
           }
         } else if (error instanceof InvalidScopeError) {