+++ examples/partial-prerendering/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @stackframe/example-partial-prerendering
 
+## 2.8.32
+
+### Patch Changes
+
+- @stackframe/stack@2.8.32
+
+## 2.8.31
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.31
+
+## 2.8.30
+
+### Patch Changes
+
+- @stackframe/stack@2.8.30
+
 ## 2.8.29
 
 ### Patch Changes