+++ packages/stack-shared/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-shared
 
+## 2.8.32
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.31
+
+### Patch Changes
+
+- Various changes
+
+## 2.8.30
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.29
 
 ### Patch Changes